requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy
pandas
snowflake-snowpark-python[pandas]
azure.keyvault
azure.identity
azure.storage.blob